backend/requirements.txt
anyio==3.7.1
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cookies==2.2.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.10
packaging==25.0
pillow==11.2.1
pipdeptree==2.26.1
PyJWT==2.9.0
python-decouple==3.8
redis==6.0.0
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
spotipy==2.25.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
youtube-search-python==1.6.5
ytmusicapi==1.10.3
